requirements.txt
pwntools==4.9.0
pycryptodome==3.17
